feat: deploy.sh

+++ scripts/deploy.sh
@@ -1,12 +1,20 @@
 #!/bin/bash
 
 APP_DIR=/home/ec2-user/app
-JAR_NAME=$(find $APP_DIR -name "*.jar" | head -n 1)
+JAR_DIR=$APP_DIR/build/libs
+JAR_NAME=$(ls $JAR_DIR/*.jar | head -n 1)
 
-echo "Using jar: $JAR_NAME"
+echo "배포 시작"
+echo "JAR = $JAR_NAME"
+
+if [ -z "$JAR_NAME" ]; then
+  echo "❌ JAR 파일을 찾지 못함"
+  exit 1
+fi
 
 PID=$(pgrep -f java)
 if [ -n "$PID" ]; then
+  echo "기존 프로세스 종료: $PID"
   kill -15 $PID
   sleep 5
 fi
